requirements.txt
# Core dependencies for YouTube API RAG application
chromadb==1.3.7
groq==0.37.1
langchain==1.1.3
langchain-chroma==1.1.0
langchain-community==0.4.1
langchain-core==1.2.0
langchain-groq==1.1.1
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.0
langgraph==1.0.5
numpy==2.3.5
pandas==2.3.3
pillow==12.0.0
python-dotenv==1.2.1
pypdf==6.4.2
requests==2.32.5
sentence-transformers==5.2.0
streamlit==1.52.1
youtube-transcript-api==1.2.3
